msm: mdss: Fix VIG pipe reset of postprocessing parameters
When a VIG pipe is destroyed or reinitialized, the PP related
parameters stored inside the pipe structure are cleared. There
was an issue with clearing the parameters. The histogram structure
was set to 0, causing a problem with histogram-related mutex and spinlock.
Rather than resetting the whole PP structure to 0, this change adds
support to only clear PP config enable flags and dirty flags. The
histogram structure is disabled but not cleared, avoiding a need to
reinitialize the histogram synchronization resources.

Change-Id: I286b528436a457b9580c3a4c2c24cc8a31dacabf
Signed-off-by: Benet Clark <benetc@codeaurora.org>